feat: webhook trigger idempotency & attempt to fix extra jobs being scheduled.
