+++ src/modules/data-ingestion/location.job.utils.ts
@@ -1,8 +1,9 @@
 import { SyncLocationsTask } from '../../trigger/data-ingestion/location/sync-locations';
+import { LogsInterface } from '../cloudshelf/logs.interface';
 import { RetailerEntity } from '../retailer/retailer.entity';
 
 export class LocationJobUtils {
-    static async schedule(retailer: RetailerEntity, reason?: string) {
+    static async schedule(retailer: RetailerEntity, reason?: string, logs?: LogsInterface) {
         const tags: string[] = [`retailer_${retailer.id}`];
         if (reason) {
             tags.push(`reason_${reason}`);
