+++ src/trigger/scheduled/safety_sync.ts
@@ -1,9 +1,9 @@
 import { FlushMode } from '@mikro-orm/core';
 import { EntityManager } from '@mikro-orm/postgresql';
+import { ProductJobUtils } from '../../modules/data-ingestion/product.job.utils';
+import { RetailerEntity } from '../../modules/retailer/retailer.entity';
 import { AppDataSource } from '../reuseables/orm';
 import { logger, schedules } from '@trigger.dev/sdk/v3';
-import { ProductJobUtils } from 'src/modules/data-ingestion/product.job.utils';
-import { RetailerEntity } from 'src/modules/retailer/retailer.entity';
 
 export const CreateSafetySyncs = schedules.task({
     id: 'create-safety-syncs',
@@ -36,7 +36,11 @@ export async function internalScheduleTriggerJobs(em: EntityManager) {
 
     for (const retailer of retailers) {
         logger.debug('Creating safety sync for retailer ' + retailer.domain);
-        await ProductJobUtils.scheduleTriggerJob(retailer, true, undefined, 'safetySync');
+        await ProductJobUtils.scheduleTriggerJob(retailer, true, undefined, 'safetySync', {
+            info: logger.info,
+            warn: logger.warn,
+            error: logger.error,
+        });
         retailer.lastSafetySyncRequested = new Date();
     }
